requirements.txt
asgiref==3.6.0
attrs==22.2.0
awsebcli==3.20.5
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blessed==1.20.0
boto3==1.26.81
botocore==1.29.81
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==39.0.2
Django==4.1.7
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
graphviz==0.20.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
MarkupSafe==2.1.2
packaging==23.0
paramiko==3.1.0
pathspec==0.10.1
pillow==9.5.0
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-gettext==4.1
pytz==2022.7.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.1.9
websocket-client==0.59.0
